NetCommHub-App/LoginHandlerTest.java
import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;

// Unit tests for the LoginHandler class
public class LoginHandlerTest {

    // The instance of the LoginHandler class to be tested
    private LoginHandler loginHandler;

    // Set up method executed before each test
    @Before
    public void setUp() {
        // Create a new instance of the LoginHandler class
        loginHandler = new LoginHandler();
    }

    // Test case to check if the login method returns a non-null array for valid input
    @Test
    public void testLoginValidInput() {
        // Call the login method and check if it returns a non-null array
        String[] result = loginHandler.login();
        assertNotNull(result);
    }

    // Test case to check if the login method handles invalid port number
    @Test
    public void testLoginInvalid() {

        // Call the login method and check if it returns null for invalid input
        assertNull(loginHandler.login());
    }
}